// src/constants/component-types.ts
export type ComponentType =
    | 'text-field'
    | 'field-array'
    | 'checkbox'
    | 'sub-form'
    | 'radio'
    | 'tabs'
    | 'tab-item'
    | 'date-picker'
    | 'time-picker'
    | 'wizard'
    | 'switch'
    | 'textarea'
    | 'select'
    | 'plain-text'
    | 'button'
    | 'input-addon-group'
    | 'input-addon-button-group'
    | 'dual-list-select'
    | 'slider'
    | 'link'
    | 'combo-box'

interface CarbonComponentTypes {
    TEXT_FIELD: 'text-field'
    TEXTAREA: 'textarea'
    SELECT: 'select'
    CHECKBOX: 'checkbox'
    SUB_FORM: 'sub-form'
    RADIO: 'radio'
    TABS: 'tabs'
    DATE_PICKER: 'date-picker'
    TIME_PICKER: 'time-picker'
    WIZARD: 'wizard'
    SWITCH: 'switch'
    PLAIN_TEXT: 'plain-text'
    FIELD_ARRAY: 'field-array'
    DUAL_LIST_SELECT: 'dual-list-select'
    SLIDER: 'slider'
    LINK: 'link'
    COMBO_BOX: 'combo-box'
    TAB_ITEM: 'tab-item'
    BUTTON: 'button'
    INPUT_ADDON_GROUP: 'input-addon-group'
    INPUT_ADDON_BUTTON_GROUP: 'input-addon-button-group'
}

declare const componentTypes: CarbonComponentTypes

export default componentTypes